1/25 20:25  메뉴얼쪽 추

+++ Coffee_iOS/Coffee_iOS/VC/반제품/HalfProductFirstVC.h
@@ -94,8 +94,7 @@
     UILabel *TotalMouthfeel_Po;
     UILabel *TotalMouthfeel_Ne;
     
-    int fix_position;
-
+    NSInteger fix_position;
 }
 
 @property (weak, nonatomic) IBOutlet JNExpandableTableView *halfFirstTableView;
