+++ Coffee_iOS/Coffee_iOS/VC/원료커핑/CoffeeFirstVC.h
@@ -69,7 +69,7 @@
     UILabel *TotalAcidity_Po;
     UILabel *TotalAcidity_Ne;
     
-    int fix_position;
+    NSInteger fix_position;
 }
 
 @property (weak, nonatomic) IBOutlet JNExpandableTableView *coffeeFirstTableView;
